requirements.txt
affine==2.4.0
alembic==1.12.0
annotated-types==0.6.0
attrs==23.1.0
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
databricks-cli==0.18.0
docker==6.1.3
entrypoints==0.4
filelock==3.12.4
Fiona==1.9.4.post1
Flask==2.3.3
fonttools==4.43.1
geopandas==0.14.0
gitdb==4.0.10
GitPython==3.1.37
greenlet==3.0.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
Mako==1.2.4
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.0
mpmath==1.3.0
networkx==3.1
numpy==1.26.0
oauthlib==3.2.2
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
planetary-computer==1.0.0
prometheus-client==0.17.1
prometheus-flask-exporter==0.22.4
protobuf==4.24.4
psutil==5.9.6
pyarrow==13.0.0
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pyparsing==3.1.1
pyproj==3.6.1
pystac==1.8.4
pystac-client==0.7.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
querystring-parser==1.2.4
rasterio==1.3.8.post1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
scikit-learn==1.3.1
scipy==1.11.3
shapely==2.0.1
six==1.16.0
smmap==5.0.1
snuggs==1.4.7
SQLAlchemy==2.0.22
sqlparse==0.4.4
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.2.0
torch==2.0.0 -f https://download.pytorch.org/whl/cpu
torchvision==0.15.1 -f https://download.pytorch.org/whl/cpu
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
waitress==2.1.2
websocket-client==1.6.4
Werkzeug==3.0.0
zipp==3.17.0
